+++ core/src/operations/mod.rs
@@ -1,0 +1,11 @@
+mod dma_inputcpy;
+mod dma_memcmp;
+mod dma_memcpy;
+mod dma_memset;
+mod profile;
+
+pub use dma_inputcpy::*;
+pub use dma_memcmp::*;
+pub use dma_memcpy::*;
+pub use dma_memset::*;
+pub use profile::*;
