+++ definitions/src/lib.rs
@@ -1,2 +1,5 @@
 mod syscall;
 pub use syscall::*;
+
+mod profile;
+pub use profile::*;
