+++ definitions/src/profile.rs
@@ -1,0 +1,4 @@
+pub const PROFILE_START_COST_ID: u16 = 1;
+pub const PROFILE_END_COST_ID: u16 = 2;
+pub const PROFILE_START_STEPS_ID: u16 = 3;
+pub const PROFILE_END_STEPS_ID: u16 = 4;
