+++ definitions/src/syscall.rs
@@ -28,3 +28,4 @@ pub const SYSCALL_DMA_MEMSET_ID: u16 = 0x816;
 pub const SYSCALL_SECP256R1_ADD_ID: u16 = 0x817;
 pub const SYSCALL_SECP256R1_DBL_ID: u16 = 0x818;
 pub const SYSCALL_BLAKE2B_ROUND_ID: u16 = 0x819;
+pub const SYSCALL_PROFILE_ID: u16 = 0x81A;
